requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-filter==0.2.1
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
hipo-drf-exceptions==0.1.3
Pillow==7.0.0
reportlab==3.5.42
